moving objects and case classes to a seperate file for organization
