+++ src/main/scala/com/microsoft/ml/spark/cognitive/TextAnalyticsSDK.scala
@@ -118,20 +118,6 @@ class TextAnalyticsLanguageDetection(override val textAnalyticsOptions: Option[T
     }
 }
 
-object DetectLanguageResponseV4 extends SparkBindings[TAResponseV4[DetectedLanguageV4]]
-
-object KeyPhraseResponseV4 extends SparkBindings[TAResponseV4[KeyphraseV4]]
-
-case class TAResponseV4[T](result: Option[T],
-                           error: Option[TAErrorV4],
-                           statistics: Option[DocumentStatistics],
-                           modelVersion: Option[String])
-
-case class TAErrorV4(errorCode: String, errorMessage: String, target: String)
-
-case class DetectedLanguageV4(name: String, iso6391Name: String, confidenceScore: Double)
-
-
 object TextAnalyticsKeyphraseExtraction extends ComplexParamsReadable[TextAnalyticsKeyphraseExtraction]
 
 class TextAnalyticsKeyphraseExtraction (override val textAnalyticsOptions: Option[TextAnalyticsRequestOptions] = None,
@@ -178,7 +164,4 @@ class TextAnalyticsKeyphraseExtraction (override val textAnalyticsOptions: Optio
       Some(ExtractKeyPhrasesResultCollection.getModelVersion))
   }
 }
-
-case class TAWarningV4 (warningCode: String, message: String)
-case class KeyphraseV4(keyPhrases: List[String], warnings: List[TAWarningV4])
